Replaced JsonObject.getDouble() calls with getInt() when applicable

+++ naga-core/src/main/java/naga/core/json/buscall/BusCallArgument.java
@@ -71,7 +71,7 @@ class BusCallArgument {
                 return new BusCallArgument(
                         json.getString(TARGET_ADDRESS_KEY),
                         JsonCodecManager.decodeFromJson(json.get(TARGET_ARGUMENT_KEY)),
-                        (int) json.getDouble(CALL_NUMBER_KEY)
+                        json.getInt(CALL_NUMBER_KEY)
                 );
             }
         };
